requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
atlassian-python-api==3.41.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cramjam==2.7.0
cryptography==38.0.4
dask==2022.12.1
Deprecated==1.2.14
EditorConfig==0.12.3
exceptiongroup==1.1.3
Faker==19.6.2
fast-query-parsers==1.0.3
fastparquet==2022.12.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.2
idna==3.4
joblib==1.3.2
jsbeautifier==1.14.9
litestar==2.1.0
locket==1.0.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgspec==0.18.2
multidict==6.0.4
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
pandas==1.5.3
partd==1.4.1
polyfactory==2.9.0
py4j==0.10.9.5
pycparser==2.21
pydantic==2.4.0
pydantic_core==2.10.0
Pygments==2.16.1
pyspark==3.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zenpy==2.0.40
